# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "티에이치스터디 API",
        "description": "티에이치스터디 API 문서",
        "contact": {
            "name": "티에이치스터디",
            "email": "admin@th-study.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "현재 서버"
        },
        {
            "url": "http://localhost:8000",
            "description": "로컬 서버"
        },
        {
            "url": "http://localhost:8080",
            "description": "도커 서버"
        },
        {
            "url": "https://www.th-study.com",
            "description": "운영 서버"
        }
    ],
    "paths": {
...
        "/api/mcp/tools/access-logs": {
            "post": {
                "tags": [
                    "MCP"
                ],
                "summary": "사람 유입 조회",
                "description": "실제 방문자의 유입 로그를 조건에 따라 조회합니다.",
                "operationId": "4083d871fe4e20ea5126a93ee3df5c48",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "group_code": {
                                        "type": "string",
                                        "example": "blog"
                                    },
                                    "categories_code": {
                                        "type": "string",
                                        "example": "develop"
                                    },
                                    "topic_code": {
                                        "type": "string",
                                        "example": "라라벨"
                                    },
                                    "user_idx": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "device_types": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "desktop",
                                                "mobile",
                                                "tablet"
                                            ]
                                        },
                                        "example": [
                                            "desktop",
                                            "mobile"
                                        ]
                                    },
                                    "ip": {
                                        "type": "string",
                                        "example": "192.168.0.1"
                                    },
                                    "referer_host": {
                                        "type": "string",
                                        "example": "www.google.com"
                                    },
                                    "access_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-27"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-27"
                                    },
                                    "has_note": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "page": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "per_page": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "사람 유입 조회 성공"
                    },
                    "401": {
                        "description": "인증 실패"
                    },
                    "422": {
                        "description": "유효성 검사 실패"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mcp/tools/bot-access-logs": {
            "post": {
                "tags": [
                    "MCP"
                ],
                "summary": "봇 유입 조회",
                "description": "검색엔진 및 AI 크롤러의 유입 로그를 조건에 따라 조회합니다.",
                "operationId": "e92d8af27868cb0e21b2324df679183b",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "group_code": {
                                        "type": "string",
                                        "example": "blog"
                                    },
                                    "categories_code": {
                                        "type": "string",
                                        "example": "develop"
                                    },
                                    "topic_code": {
                                        "type": "string",
                                        "example": "라라벨"
                                    },
                                    "bot_name": {
                                        "type": "string",
                                        "example": "Googlebot"
                                    },
                                    "access_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-27"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-27"
                                    },
                                    "referer_host": {
                                        "type": "string",
                                        "example": "www.google.com"
                                    },
                                    "has_note": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "page": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "per_page": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "봇 유입 조회 성공"
                    },
                    "401": {
                        "description": "인증 실패"
                    },
                    "422": {
                        "description": "유효성 검사 실패"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mcp/tools/conversion-logs": {
            "post": {
                "tags": [
                    "MCP"
                ],
                "summary": "유입 후 전환 조회",
                "description": "사용자가 유입된 이후 발생한 클릭 및 페이지 이동 등의 전환 로그를 조회합니다.",
                "operationId": "316c6ba76d2bfcfde16a4501ad939aee",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "group_code": {
                                        "type": "string",
                                        "example": "blog"
                                    },
                                    "categories_code": {
                                        "type": "string",
                                        "example": "develop"
                                    },
                                    "topic_code": {
                                        "type": "string",
                                        "example": "라라벨"
                                    },
                                    "user_idx": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "device_types": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "desktop",
                                                "mobile",
                                                "tablet"
                                            ]
                                        },
                                        "example": [
                                            "desktop",
                                            "mobile"
                                        ]
                                    },
                                    "ip": {
                                        "type": "string",
                                        "example": "192.168.0.1"
                                    },
                                    "referer_host": {
                                        "type": "string",
                                        "example": "www.google.com"
                                    },
                                    "conversion_type": {
                                        "type": "string",
                                        "example": "outbound"
                                    },
                                    "conversion_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-27"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-27"
                                    },
                                    "has_note": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "page": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "per_page": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "유입 후 전환 조회 성공"
                    },
                    "401": {
                        "description": "인증 실패"
                    },
                    "422": {
                        "description": "유효성 검사 실패"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}